Add compatability for new Gmail redesign
Gmail redesign caused issues with the email ID. Location slug no longer corresponds to the email ID. The email ID or thread ID seems to be accessed by a number of other plugins from the `h2[data-legacy-thread-id]` element so this update pulls from that element as it's present in both new and classic Gmail.

ServiceLink is also now inserted as an adjacent element to the span as the new redesign was repainting that element after the service link was being added.